+++ src/Service/MediaConvert/src/ValueObject/DashIsoGroupSettings.php
@@ -51,6 +51,19 @@ final class DashIsoGroupSettings
     private $baseUrl;
 
     /**
+     * Specify whether MediaConvert generates I-frame only video segments for DASH trick play, also known as trick mode.
+     * When specified, the I-frame only video segments are included within an additional AdaptationSet in your DASH output
+     * manifest. To generate I-frame only video segments: Enter a name as a text string, up to 256 character long. This name
+     * is appended to the end of this output group's base filename, that you specify as part of your destination URI, and
+     * used for the I-frame only video segment files. You may also include format identifiers. For more information, see:
+     * https://docs.aws.amazon.com/mediaconvert/latest/ug/using-variables-in-your-job-settings.html#using-settings-variables-with-streaming-outputs
+     * To not generate I-frame only video segments: Leave blank.
+     *
+     * @var string|null
+     */
+    private $dashIframeTrickPlayNameModifier;
+
+    /**
      * Specify how MediaConvert writes SegmentTimeline in your output DASH manifest. To write a SegmentTimeline in each
      * video Representation: Keep the default value, Basic. To write a common SegmentTimeline in the video AdaptationSet:
      * Choose Compact. Note that MediaConvert will still write a SegmentTimeline in any Representation that does not share a
@@ -226,6 +239,7 @@ final class DashIsoGroupSettings
      *   AdditionalManifests?: null|array<DashAdditionalManifest|array>,
      *   AudioChannelConfigSchemeIdUri?: null|DashIsoGroupAudioChannelConfigSchemeIdUri::*,
      *   BaseUrl?: null|string,
+     *   DashIFrameTrickPlayNameModifier?: null|string,
      *   DashManifestStyle?: null|DashManifestStyle::*,
      *   Destination?: null|string,
      *   DestinationSettings?: null|DestinationSettings|array,
@@ -251,6 +265,7 @@ final class DashIsoGroupSettings
         $this->additionalManifests = isset($input['AdditionalManifests']) ? array_map([DashAdditionalManifest::class, 'create'], $input['AdditionalManifests']) : null;
         $this->audioChannelConfigSchemeIdUri = $input['AudioChannelConfigSchemeIdUri'] ?? null;
         $this->baseUrl = $input['BaseUrl'] ?? null;
+        $this->dashIframeTrickPlayNameModifier = $input['DashIFrameTrickPlayNameModifier'] ?? null;
         $this->dashManifestStyle = $input['DashManifestStyle'] ?? null;
         $this->destination = $input['Destination'] ?? null;
         $this->destinationSettings = isset($input['DestinationSettings']) ? DestinationSettings::create($input['DestinationSettings']) : null;
@@ -276,6 +291,7 @@ final class DashIsoGroupSettings
      *   AdditionalManifests?: null|array<DashAdditionalManifest|array>,
      *   AudioChannelConfigSchemeIdUri?: null|DashIsoGroupAudioChannelConfigSchemeIdUri::*,
      *   BaseUrl?: null|string,
+     *   DashIFrameTrickPlayNameModifier?: null|string,
      *   DashManifestStyle?: null|DashManifestStyle::*,
      *   Destination?: null|string,
      *   DestinationSettings?: null|DestinationSettings|array,
@@ -320,6 +336,11 @@ final class DashIsoGroupSettings
     public function getBaseUrl(): ?string
     {
         return $this->baseUrl;
+    }
+
+    public function getDashIframeTrickPlayNameModifier(): ?string
+    {
+        return $this->dashIframeTrickPlayNameModifier;
     }
 
     /**
@@ -464,6 +485,9 @@ final class DashIsoGroupSettings
         }
         if (null !== $v = $this->baseUrl) {
             $payload['baseUrl'] = $v;
+        }
+        if (null !== $v = $this->dashIframeTrickPlayNameModifier) {
+            $payload['dashIFrameTrickPlayNameModifier'] = $v;
         }
         if (null !== $v = $this->dashManifestStyle) {
             if (!DashManifestStyle::exists($v)) {